# src/main/java/au/com/cap/ref/transactions/genreport/api/GenerateSummaryReportApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.4.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package au.com.cap.ref.transactions.genreport.api;

import au.com.cap.ref.transactions.genreport.model.DailySummaryReport;
import au.com.cap.ref.transactions.genreport.model.ErrorResponse;
import org.springframework.core.io.Resource;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2020-03-22T22:26:15.976+11:00")

@Api(value = "generateSummaryReport", description = "the generateSummaryReport API")
public interface GenerateSummaryReportApi {

    @ApiOperation(value = "", nickname = "generateSummaryReport", notes = "Generates a daily summary report for the given feed (.txt) file", response = DailySummaryReport.class, tags={ "Summary Report", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "OK", response = DailySummaryReport.class),
        @ApiResponse(code = 400, message = "__Bad Request__ Possible Errors   | Error Code | Error Message | Error Description |   | ------- | --------- | ------ |   | 1001   | Invalid Input Parameter | The given {parameterName}-{parametersValue} is missing or invalid | ", response = ErrorResponse.class),
        @ApiResponse(code = 500, message = "__Internal Server Error__ Possible Errors   | Error Code | Error Message | Error Description |   | -------- | --------- | --------|   | 10000   | Technical Error | {Error Details} | ", response = ErrorResponse.class) })
    @RequestMapping(value = "/generateSummaryReport",
        produces = { "application/json", "multipart/form-data" }, 
        consumes = { "multipart/form-data" },
        method = RequestMethod.POST)
    ResponseEntity<DailySummaryReport> generateSummaryReport(@ApiParam(value = "Globally unique message identifier - GUID. This uniquely identifies the call" ,required=true) @RequestHeader(value="x-messageId", required=true) String xMessageId,@ApiParam(value = "Consumer generated message identifier. This groups together number of API calls comprising a business transaction" ,required=true) @RequestHeader(value="x-appCorrelationId", required=true) String xAppCorrelationId,@ApiParam(value = "Unique identifier for the user invoking the API (e.g. Staff Id)" ,required=true) @RequestHeader(value="x-userId", required=true) String xUserId,@ApiParam(value = "file detail") @Valid @RequestPart("file") MultipartFile inputFile);

}